pyproject.toml
[project]
name = "bike-parking"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow<3.0.0",
    "dbt-postgres>=1.10.0",
    "fastapi[standard]>=0.135.2",
    "jinja2>=3.1.6",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.33.0",
    "uvicorn>=0.42.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "ipykernel>=7.2.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "pandas>=3.0.2",
    "pytest>=9.0.2",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "sqlalchemy>=1.4.54",
]
